requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
cryptography==41.0.5
python-multipart==0.0.6